requirements.txt
aiohttp>=3.9.0
aiogram>=3.15.0
apscheduler>=3.10.0,<4.0.0
groq>=0.9.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
requests>=2.31.0
